pipeline.py
"""
requirements: httpx, pydantic, beautifulsoup4, asyncio, urllib3, logging
"""

from typing import List, Union, Generator, Iterator, Dict, Optional
from pydantic import BaseModel, Field
import asyncio
import logging
from urllib.parse import quote_plus

import httpx
from bs4 import BeautifulSoup

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class ValidationResult(BaseModel):
    """Модель для хранения результатов валидации документа"""
    source_statuses: Dict[str, str] = Field(alias="источник_статусы")
    status: str = Field(alias="статус")
    confidence: float = Field(alias="уверенность")

class DocumentReference(BaseModel):
    """Модель для хранения информации о документе"""
    raw: str
    type: Optional[str] = None
    number: Optional[str] = None
    date: Optional[str] = None
    title: Optional[str] = None

class Pipeline:
    
    class Valves(BaseModel):
        '''
        Класс для хранения параметров пайплайна, которые будут задаваться внешней настройкой
        '''
        max_sources: Optional[int] = 10
        request_timeout: float = 10.0
        batch_timeout: float = 30.0
        max_connections: int = 10
        max_keepalive: int = 5
        user_agent: str = "Mozilla/5.0 (compatible; DocumentValidator/1.0; +https://example.com)"
        
    def __init__(self):
        """Инициализация пайплайна с настройкой источников проверки"""
        self.name = "Document Validation Pipeline"
        self.valves = self.Valves()
        
        # Источники для проверки
        self.sources = {
            # законодательство
            "pravo.gov.ru": "https://pravo.gov.ru/search/?query={query}",
            "consultant.ru": "https://www.consultant.ru/search/?q={query}",
            "garant.ru": "https://www.garant.ru/search/?q={query}",
            "rulaws.ru": "https://rulaws.ru/search/?q={query}",
            # стандарты
            "docs.cntd.ru": "https://docs.cntd.ru/search/?searchtext={query}",
            "gostrf.com": "https://gostrf.com/search/?q={query}",
            "standartgost.ru": "https://standartgost.ru/search/?q={query}",
            "gov.spb.ru": "https://www.gov.spb.ru/search/?q={query}",
            "gostassistent.ru": "https://gostassistent.ru/search?q={query}",
        }

        # Ключевые слова для определения статуса
        self.expired_keywords = [
            "не действует", "утратил силу", "утратила силу",
            "утратило силу", "признан утратившим силу",
            "отменен", "отменена", "заменен", "заменена",
        ]
        self.active_keywords = [
            "действует", "действующая редакция", "актуально",
        ]

    async def _fetch(self, client: httpx.AsyncClient, url: str) -> tuple[str, str]:
        """Выполняет запрос к источнику и возвращает результат"""
        try:
            logger.info(f"Проверка источника: {url}")
            resp = await client.get(url, timeout=self.valves.request_timeout)
            resp.raise_for_status()
            logger.info(f"Успешно проверен {url}")
            return url, resp.text.lower()
        except httpx.TimeoutException:
            logger.warning(f"Таймаут при проверке {url}")
            return url, ""
        except httpx.HTTPStatusError as e:
            logger.warning(f"HTTP ошибка {e.response.status_code} при проверке {url}")
            return url, ""
        except Exception as e:
            logger.warning(f"Ошибка при проверке {url}: {str(e)}")
            return url, ""

    async def validate_reference(self, ref: Union[str, DocumentReference]) -> ValidationResult:
        """Проверяет документ по всем источникам"""
        if isinstance(ref, DocumentReference):
            query = quote_plus(ref.raw)
        else:
            query = quote_plus(ref)

        headers = {"User-Agent": self.valves.user_agent}
        limits = httpx.Limits(
            max_keepalive_connections=self.valves.max_keepalive,
            max_connections=self.valves.max_connections
        )

        tasks = []
        async with httpx.AsyncClient(
            headers=headers,
            follow_redirects=True,
            limits=limits,
            timeout=self.valves.request_timeout
        ) as client:
            logger.info(f"Начало валидации для: {query}")
            for i, (name, tmpl) in enumerate(self.sources.items()):
                if self.valves.max_sources and i >= self.valves.max_sources:
                    break
                url = tmpl.format(query=query)
                tasks.append(self._fetch(client, url))

            try:
                results = await asyncio.wait_for(
                    asyncio.gather(*tasks),
                    timeout=self.valves.batch_timeout
                )
            except asyncio.TimeoutError:
                logger.error("Таймаут при сборе результатов")
                results = [(url, "") for url in self.sources.values()]

        status_counts = {"expired": 0, "active": 0, "unknown": 0}
        source_status: Dict[str, str] = {}

        for url, html in results:
            if not html:
                source_status[url] = "unknown"
                status_counts["unknown"] += 1
                continue

            expired = any(k in html for k in self.expired_keywords)
            active = any(k in html for k in self.active_keywords)
            
            if expired and not active:
                st = "просрочено"
                status_counts["expired"] += 1
            elif active and not expired:
                st = "действительно"
                status_counts["active"] += 1
            else:
                st = "неизвестно"
                status_counts["unknown"] += 1
            source_status[url] = st

        total = sum(status_counts.values())
        if total == 0:
            final = "Неизвестно"
            conf = 0.0
        elif status_counts["expired"] >= status_counts["active"] and status_counts["expired"] > 0:
            final = "Просрочено"
            conf = status_counts["expired"] / total
        elif status_counts["active"] > status_counts["expired"]:
            final = "Действительно"
            conf = status_counts["active"] / total
        else:
            final = "Неизвестно"
            conf = 0.0

        return ValidationResult(
            источник_статусы=source_status,
            статус=final,
            уверенность=round(conf, 2)
        )

    def extract_references(self, text: str) -> List[DocumentReference]:
        """Извлекает ссылки на документы из текста"""
        # TODO: Implement reference extraction logic
        # This is a placeholder that should be implemented based on your needs
        return [DocumentReference(raw=text)]

    async def pipe(
        self,
        user_message: dict,
        model_id: str,
        messages: List[dict],
        body: dict
    ) -> Union[str, Generator, Iterator]:
        """
        Основной метод обработки запросов
        
        Args:
            user_message: Сообщение пользователя
            model_id: Идентификатор модели
            messages: История сообщений
            body: Дополнительные параметры запроса
        
        Returns:
            Union[str, Generator, Iterator]: Результат обработки запроса
        """
        try:
            # Извлекаем текст из сообщения пользователя
            content = user_message.get("content", "")
            
            # Извлекаем ссылки на документы из текста
            references = self.extract_references(content)
            
            # Проверяем каждую ссылку
            results = []
            for ref in references:
                result = await self.validate_reference(ref)
                results.append({
                    "reference": ref.dict(),
                    "validation": result.dict()
                })
            
            return {
                "status": "success",
                "results": results
            }
            
        except Exception as e:
            logger.error(f"Ошибка в пайплайне: {str(e)}")
            return {
                "status": "error",
                "message": str(e)
            } 