requirements.txt
python-docx~=1.1
pandas~=2.2
openpyxl~=3.1
httpx[socks]==0.25.2
beautifulsoup4~=4.12
lxml~=5.2
regex~=2024.4
mistralai~=0.0.10  # official mistral api client 
typer[all]~=0.12
rich~=13.7
selenium~=4.15
webdriver-manager~=4.0
requests>=2.28.0
aiohttp>=3.8.0
# redis>=4.0.0  # Убрано - используем локальный кеш
beautifulsoup4>=4.11.0
lxml>=4.9.0
click>=8.0.0
rich>=12.0.0 